+++ requests/employee.http
@@ -1,0 +1,2 @@
+### List products
+GET http://localhost:3000/employee
