--- requests/products.http
@@ -1,2 +1,0 @@
-### List products
-GET http://localhost:3000/products
