+++ src/Employee/EmployeeService.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+import ProductRepository from './EmployeeRepository';
+
+@Injectable()
+export class EmployeeService {
+	constructor(private readonly productRepository: ProductRepository) {}
+
+	async listEmployees() {
+		return this.productRepository.getAllEmployees();
+	}
+}
