+++ src/app.module.ts
@@ -4,16 +4,16 @@ import { User } from './User/User';
 import { UserModule } from './User/UserModule';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ProductsController } from './Product/ProductsController';
-import { ProductsModule } from './Product/ProductModule';
-import { Product } from './Product/Product';
+import { ProductsController } from './Employee/EmployeeController';
+import { ProductsModule } from './Employee/EmployeeModule';
+import { Employee } from './Employee/Employee';
 
 @Module({
 	imports: [
 		TypeOrmModule.forRoot({
 			type: 'sqlite',
 			database: 'db.sqlite',
-			entities: [User, Product],
+			entities: [User, Employee],
 			synchronize: true,
 			logging: true,
 		}),
