+++ test/Products/products.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ProductsController } from '../../src/Product/ProductsController';
+import { ProductsController } from '../../src/Employee/EmployeeController';
 
 describe('ProductsController', () => {
 	let controller: ProductsController;
